requirements (1).txt
langchain
langchain-groq
langchain-community
langgraph
duckduckgo-search
rich
python-dotenv
requests
ddgs